requirements.txt
Django>=5.0,<6.1
psycopg[binary]>=3.1
djangorestframework>=3.15
django-filter>=24.2
